requirements.txt
beautifulsoup4>=4.14.2
python-dateutil>=2.9.0.post0
requests>=2.32.5
